Updates to build and link on Solaris
config/platform.mk.template:
Don't use the -Wno-write-strings compile flag on Solaris.
For now, we're still using the Solaris compiler
(/opt/SUNWspro/bin/cc), not g++, so it doesn't have that.

lib/core/src/rcMisc.c:
On Solaris, use putenv instead of setenv as putenv is no longer
available (at least on multivac.sdsc.edu, which is SunOS 5.9).  

Note that I needed to make this change even in iRODS 2.5 to build on
Multivac to avoid the missing external on setenv, so it looks like it
is a difference for this version of Solaris.


git-svn-id: svn://irodssvn.ucsd.edu/trunk@4120 ea137bd6-aa86-47b6-b90d-b101d9f19a90