In chlGeneralUpdate, perform a rollback after an error so that the transaction is closed out and subsequent updates can succeed.  The errors returned were CAT_SQL_ERROR which were due to "Error while executing the query (non-fatal); ERROR: current transaction is aborted, commands ignored until end of transaction block" at the DBMS/ODBC level.
This fix was developed solve a problem with repeated attempted updates
when a row may already exist noticed by Alinga (see irods-chat).  This
seems to solve the problem.  This could be fixed via a call to
chlRollback at the caller's level but for a function like this it is
probably better to handle it here.


git-svn-id: svn://irodssvn.ucsd.edu/trunk@4509 ea137bd6-aa86-47b6-b90d-b101d9f19a90