First set of additions for ticket-based-access including an preliminary iticket command and ICAT operations to update ticket tables, new table definitions, some defines and setup for general-query, new tests for the new SQL (in icatTest.pl for 'irodsctl devtest'), iget/library additions to pass in the ticket, etc.
The small changes in iget.c, dataObjOpr.c, rodsKeyWdDef.h, and
getUtil.c allow the iget command to pass in the ticket via '-t
ticket-string' (we already had -t reserved as a standard option for
tickets).  This arrives in the general-query (icatGeneralQuery.c)
which for now just logs the ticket.

The rest (iticket, ICAT additions, etc) just creates and modifies
tickets with various options.  I still need to add logic in the ICAT
to accept and check them, restrict the displaying of the
ticket-strings, handle 'write' tickets, etc.

This step uses rcUserAdmin for the client/server interaction but this
will be changed to a new client/server call in a later step.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@4513 ea137bd6-aa86-47b6-b90d-b101d9f19a90