More support for tickets: read-access tickets by user anonymous and strict ACL mode and read tickets on collections.
Changed the system to transfer the ticket before other interactions
(queries) to simplify changes on the both client and server/ICAT side.
For now, this is also via an existing client/server call (rcUserAdmin)
but will be changed to a new one later.  Also for now, the ticket is
still passed in with some queries, but I will back this out if I stick
with this new method.

Made changes/additions so that 'read' tickets work for 'anonymous'
user and/or with when running in ACL 'strict' mode.  This is handled
via additional SQL checks in general-query logic using the
pre-transferred ticket.

Made changes/additions so that 'read' tickets on collections are also
now supported.  This also involved changes in some general-query SQL
using the session ticket.

Fixed a couple bugs in the previously committed ticket code.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@4559 ea137bd6-aa86-47b6-b90d-b101d9f19a90