Add ticket optional restrictions to groups and handle client host address better.
Add the ticket feature to be able to restrict tickets to certain
groups.  Add a new table, ticket modification logic, general-query
fields, 'iticket ls' display, check group membership when a ticket is
used, add a new error code and new tests in icatTest.pl to cover the
new SQL forms.

For the ticket client host address (for host restrictions), the old
logic only worked when the client was connecting directly to the IES.
The new logic, currently in rsUserAdmin, on a non-IES, will provide
the client host addresss in the ticket session message and the
chlModTicket function, for session setup, will use it when provided.
The old logic is still partially in place which I will clean up after
I am sure this works fine in all cases.


git-svn-id: svn://irodssvn.ucsd.edu/trunk@4576 ea137bd6-aa86-47b6-b90d-b101d9f19a90