More support for write tickets; this is largely working now so I want to get it into svn, but it still needs more testing and development.
Added ICAT code to check and allow the use of write-tickets in more
cases, and to check the various restrictions.  Also added code to
track the number of bytes written via write tickets, if a limit is in
force.  Added iticket sub-commands and ICAT support to set write-byte
or write-file limits on tickets.  Added general-query definitions for
these and use them for 'iticket ls'.  Added error codes for these
limits.

Note that a write ticket can be for a collection or an individual file
(data-object) so there are many cases involved.

Due to how the access controls work, there are somes cases where this
behaves differently than one might expect.  In some situations, when
write-ticket uses are being counted, one use can be counted as two
(internally, there are two control points).  If the file exists and is
written as same size, that size will not be counted (even if there is
a limit on write bytes) because no modification of the file size is
being performed by the higher-level (not-ICAT) functions.  If the file
exists and is written, the owner remains as before.  If a new file is
being added, it will be owned by the user writing it.  When a
write-bytes limit is in force, a file that causes the count to exceed
that limit will be allowed (enforcing this at a later point causes a
0-length file to be inserted, so it is allowed; but the next file will
be disallowed).

Also fix a compile problem in iticket.c on Solaris (C compiler)
noticed by Jean-Yves.  g++ allows it but an 'and' should be '&&'.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@4634 ea137bd6-aa86-47b6-b90d-b101d9f19a90