Add a client/server call and ICAT support to allow an admin user to create a temporary one-time password for another user, which can then be used to log in as that user.  This is needed by the iPlant project where their web interface needs to spawn a sub-process to operate as the authenticated (at their level) user.
This call, rcGetTempPasswordForOther, is similar to the
rcGetTempPassword except it is privileged and has an input argument,
which is the user to create the password for.

The existing ICAT chlMakeTempPw function required some minor changes
to accept and use optional input 'otherUser' (if the user is admin).
As with rcGetTempPassword the result is returned as a value that needs
to be hashed with the user's password to form the actual password (to
keep it concealed on the network).  In this case, the user's password
to hash with is the caller's (not otherUser's), of course.

A sequence was added to icatMiscTest.pl (which is part of 'devtest')
using an updated test_chl.c to test this.  It seems to work fine.
Other than this test, the 'C' clients (i-commands) do not make use of
this.


git-svn-id: svn://irodssvn.ucsd.edu/trunk@4638 ea137bd6-aa86-47b6-b90d-b101d9f19a90