More updates for tickets.
iticket.c:
Add help text to explain setting/clearing group restrictions on
tickets (this was already working, but not described).
Add to the help text to say that the input host for doing
host-restrictions can be either an host name (DNS) or IP address.
Remove some unneeded code.

icatHighLevelRoutines.h and icatHighLevelRoutines.c:
In chlModTicket, when adding or removing a host restriction, convert
the input string (host or IP) to an IP address (if the input is an IP
the lookup returns the IP address).

rodsErrorTable.h and rodsLog.c:
Add a new error code for this.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@4721 ea137bd6-aa86-47b6-b90d-b101d9f19a90