More updates for tickets.
icatGeneralQuery.c:
Fix a problem in tacket-based access control in 'strict' mode and/or
for user 'anonymous'.  If there's already SQL checking for a data or
collection ticket, do not add another one on collections (would not
allow access when it sometimes should).

iticket.c:
Changes to handle listing tickets when the target collection or
data-object no longer exists (the normal query won't find them because
there's no matching collection or data-object rows).



git-svn-id: svn://irodssvn.ucsd.edu/trunk@4728 ea137bd6-aa86-47b6-b90d-b101d9f19a90