Add audit calls for tickets: create, modify, delete, and use.
When a ticket is created, two audit calls are made, one to record the
ticketString and another to record the ticket target (data-object or
collection).

When a ticket is used, there are also two audit calls, when the
initial ticket-string is received and at a point in its use in the
ICAT mid layer (cml).  The first makes sure ticket uses are always
audited (altho it will record failures too) and the second has the
ticket object-id so it is recorded too (it's not available at the
first call so this seems to be the best solution).  



git-svn-id: svn://irodssvn.ucsd.edu/trunk@4731 ea137bd6-aa86-47b6-b90d-b101d9f19a90