Prevent sensitive files from being writable by 'other' even if the user building irods has a 'umask' set to be unrestricted (e.g. umask 0).
In 'irodssetup', check the umask and if open, set it to be more
appropriate and warn the user.  In 'configure', also set the umask if
needed (having been set in irodssetup tho, this will only be needed if
run manually).  In the main Makefile, do a recursive chmod to make
sure files are not writeable by 'other' (again, not needed under
'irodssetup' but only if 'make' is run manually and user's umask is
open).


git-svn-id: svn://irodssvn.ucsd.edu/trunk@4782 ea137bd6-aa86-47b6-b90d-b101d9f19a90