More improvements for previous commit for handlinge user umask 0 or similar.
In irodssetup, if the umask is open, run ichmod to remove other-write
access to the distribution, as this is needed in some cases.

In the main Makefile, add a use of 'make chmod' for the 'test' target
so that it will be run via irodssetup (setup does a build of each
target, not just a 'make').




git-svn-id: svn://irodssvn.ucsd.edu/trunk@4783 ea137bd6-aa86-47b6-b90d-b101d9f19a90