Fix a bug in ICAT user-defined meta-data logic that was preventing the reuse of some AVU definitions when the units were not used (the AVU definitions are rows in r_meta_main).
At least for Postgres, and for how the rows are currently added, a SQL
test for "IS NULL" is not enough, it needs check on it being equal to
a null string, as the unit column when unused isn't 'NULL' (unknown
status to the DBMS) but is a null string.

This logic has been like this at least since 2.4.1 but the bug was
mostly harmless, just adding an extra row in r_meta_main when it
wasn't needed, reducing performance some in some cases, and not fully
utilizing the ICAT schema as designed.  All the user-defined meta-data
operations appear to work fine even with this bug, which is why it
hasn't been noticed in tests or in production.

The bug was noticed when Thomas Ledoux was developing a feature to
allow another AVU operation (set) and it took a while for him and I to
realize that this was happening.  He was seeing AVUs that were
unshared, whereas I expected them to be shared.  Now that we have this
resolved, we can evaluate his revised meta 'set' mods for possible
inclusion into svn.  They need to handle both shared and unshared
AVUs.


git-svn-id: svn://irodssvn.ucsd.edu/trunk@4827 ea137bd6-aa86-47b6-b90d-b101d9f19a90