Add a 'iadmin modrescdatapath' capability to update data-obj paths. This is sometimes needed after a resource path is updated (if a resource has been moved).  This addresses a problem recently encountered by Leesa, which she resolved via direct SQL. See the 'iadmin h modrescdatapath' text for more information.
This is different than most commands in that it is relying on the
admin-user to input strings that will be inserted into tables, so it
warns about this and prompts to confirm.  So operation is somewhat
dangerous, but better than having them use SQL.

Most of this update is adding a sub-command to iadmin and a new ICAT
function, a link to it, and other ICAT functionality.

icatLowLevelOdbc.c and icatLowLevelOracle.c were modified so that the
row count of updated rows can be returned for these types of calls.

icatTest.pl was updated to include tests of the new functionality so
the tinderbox tests (devtest) can still succeed, exercising the new
SQL.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@4923 ea137bd6-aa86-47b6-b90d-b101d9f19a90