Bug fix: could not retrieve dataObjInfo for files with no filesystem meta-data
If the FILESYSTEM_META support was enabled, the dataObjInfo could not be retrieved for data objects with no entries in the r_objt_filesystem_meta table. This could pose a problem for systems where they turned on this support but already had data in iRODS. Changed getDataObjInfo() to retrieve file system meta-data in a separate query, and to ignore it if not found.

Signed-off-by: Chris Smith <chris@distributedbio.com>

git-svn-id: svn://irodssvn.ucsd.edu/trunk@4963 ea137bd6-aa86-47b6-b90d-b101d9f19a90