First phase for PAM authentication; new API client/server call, ICAT code, client changes, new error codes.
The iRODS server creates a new iRODS password (a random string) which
the i-commands then use.  Only 'iinit' makes the new PAM
authentication call and then stores the generated password into the
auth file (.irodsA usually), and other i-commands perform the regular
irods authentication exchange.  This protects the PAM (system)
passwords better as they are only used during the initial
authentication and then iRODS-specific passwords (credentials) are
used.  No system passwords are stored in .irodsA files (the contents
of which are obfuscated but not encrypted).

These generated passwords (iRODS-PAM passwords) expire after two weeks
(this can be changed via a #define).  If the user does another 'iinit'
(from any host), the existing generated (iRODS-PAM) password's life is
extended.  It doesn't generate a new password since a different host
might then have a non-matching .irodsA contents.  Since these are
based on the PAM credentials (logically) we don't want them to last
too long but they need to be long enough for batch jobs, etc.

The -V option on iinit displays the iRODS-PAM password.

This is just the first part of this.  What remains to be done
includes:

1) Actually call the PAM authentication process on the server side
(based on PamAuthCheck.c committed in revision 4919).  For testing now
it bypasses this and for this commit, it just always returns an error.

2) Encrypt the new API function exchange.  We are planning to use
OpenSSL encryption for this, which will provide strong, well-tested,
encryption and also authenticate the server (via certificates) since,
for a system password in particular, the client needs to confirm that
it is talking with a legitimate iRODS server.

3) Add tests, in particular for the new SQL.  This may require some
changes on the server side and some new test code.  For now, the log
messages use "xSQL" to bypass checks in devtest and so tinderbox.

4) Add an iadmin command to remove (or expire) these iRODS-PAM
passwords for specified users.  This is needed when admins change
users' system passwords and want to require re-authentication for
iRODS access too.


git-svn-id: svn://irodssvn.ucsd.edu/trunk@4964 ea137bd6-aa86-47b6-b90d-b101d9f19a90