Preliminary support for running PAM auth over SSL
This commit contains the preliminary support for doing the PAM authentication over an SSL secured socket. Since PAM requires a plain text password to be passed to the server, SSL is used to encrypt the stream to secure the transfer.

There are two client-driven routines for this: sslStart and sslEnd. Before calling rcPamAuthRequest(), sslStart() is called to set up SSL between the client and server on the already connected socket. Once rcPamAuthRequest() returns, sslEnd() tears down the SSL connection, leaving the connected socket in place for subsequent (normal) communications.

There are still a number of limitations and this hasn't been fully tested, so only turn on USE_SSL in config.mk if you enjoy living on the edge. :-)

Signed-off-by: Chris Smith <chris@distributedbio.com>

git-svn-id: svn://irodssvn.ucsd.edu/trunk@4973 ea137bd6-aa86-47b6-b90d-b101d9f19a90