Allow relocation of default server paths
Two parts to this change. The first is to define the irodsStateDir environment variable and DEF_STATE_DIR define as a location for packedRei and lockFileDir directories (needed for linux package creation according to lsb). The second is to allow overriding of DEF_CONFIG_DIR, DEF_STATE_DIR and DEF_LOG_DIR at compile time.

Signed-off-by: Chris Smith <chris@distributedbio.com>

git-svn-id: svn://irodssvn.ucsd.edu/trunk@4997 ea137bd6-aa86-47b6-b90d-b101d9f19a90