Add tests for the SQL performed for PAM authentication.  Remove the temporary 'commenting-out' (xSQL) so that these tests are required for 'devtest' to succeed.
Add anther test function in chl_test.c for this, add tests in
icatMiscTest.pl, and make some changes in the ICAT code to support the
tests.  As part of the ICAT changes, remove an ifdef PAM_AUTH
requirement in chlCheckAuth.  This is fine as it is almost no overhead
(basically a string compare) and then the tests can work without
PAM_AUTH enabled.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@5003 ea137bd6-aa86-47b6-b90d-b101d9f19a90