Implement an option to perform upper-case in queries and use it in imeta and iquest, providing a basic case-insensitive capability.  This was requested recently on irods-chat in "Case-insensitive AVU queries (Jargon)".
Added a general option to general-query to request the feature. When
present this causes the ICAT general-query functions to wrap the
table.column names in the 'where' cause with 'upper( )'.  This has
been tested for Postgres and should work from MySQL but may not
(currently) for Oracle.

In 'imeta', added a 'upper' command to toggle between upper-case mode
and, wnen on, pass the option to general-query for the 'qu' (query)
sub-command.  Note that since all columns in 'where' get 'upper( )'
added, this applies to the AVU Attribute names as well as the Value
and Units.

In 'iquest' added an option, 'uppercase' (actually,
'uppper[anything]'), to request this.  A test and example using a
case-insensitive data-object name matching seems to work fine.

Added a basic test to icatTest.pl (used in 'devtest').


git-svn-id: svn://irodssvn.ucsd.edu/trunk@5069 ea137bd6-aa86-47b6-b90d-b101d9f19a90