Fix some problems found by 'devtest' when using Oracle Xe (on srbbrick15) as the ICAT DB, found by 'devtest'.
icatHighLevelRoutines.c: 

1) Avoid an extra bind variable.  Apparently ODBC doesn't mind having
additional ones available but OCI noticed the error (and returned a
failure).  Not sure if this is just for Oracle/Xe, or if it now occurs
with other Oracle versions too.

2) Add a 'distinct' in a quota calculation SQL as there are sometimes
duplicate rows in the r_quota_usage table.  Not sure if ODBC/Postgres
doesn't encounter the duplicates or if the SQL form used doesn't have
a problem with duplicates.

icatLowLevelOracle.c:

1) In cllGetRowCount, return noResultRowCount if the input statement
number is negative (i.e. none), matching what is done in the ODBC
version.  Not sure why this wasn't a problem before (or if it was) but
this was causing a segfault in when processing the 'iadmin
modrescdatapaths' request.  Perhaps 'modrescdatapaths' hadn't been
tested with Oracle since it was added.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@5070 ea137bd6-aa86-47b6-b90d-b101d9f19a90