For 'ils -A' (access) on collections, have it display user-groups but no longer also expand user-groups to individual users, so this now matches how it is done for data-objects.  This is needed by a large iRODS site (iPlant) so they can use group ACLs and avoid very large numbers of user ACLs which seems to be causing a performance problem.
This is done via a new specific-query.  If the specific-query fails,
it will use the previous general-query call as a fall-back, but that
will expand the groups to users.

The specific-query API function has been expanded to include a
zoneHint to operate properly in a federated environment.  This is a
new condInput in the specific-query input structure, so the zoneHint
can be used so the server (rs) function can connect to remote zones
when needed (since this will be used for any ils -A on a collection).
Note that this changes the protocol for the specific-query so clients
will need to be in sync with the server.  Since, until now,
specific-query is only lightly used, this should be an easy transition
for most sites.

For this to work properly, a new specific query needs to be defined.
When iRODS is installed from scratch, the table will be set with the
proper value.  To do this interactively, you can use the iadmin asp
(add specific query) command:
asq "select distinct R_USER_MAIN.user_name ,R_USER_MAIN.zone_name, R_TOKN_MAIN.t
oken_name from R_USER_MAIN , R_TOKN_MAIN, R_OBJT_ACCESS, R_COLL_MAIN where R_OBJ
T_ACCESS.object_id = R_COLL_MAIN.coll_id AND r_COLL_MAIN.coll_name = ? AND R_TOK
N_MAIN.token_namespace = 'access_type' AND R_USER_MAIN.user_id = R_OBJT_ACCESS.u
ser_id AND R_OBJT_ACCESS.access_type_id = R_TOKN_MAIN.token_id" ShowCollAcls

(put the above asq line (as one line) into a file and run 'iadmin < file')

The general-query expands the group names into individual users due to
an artifact of how the tables are defined that allow it to generate
SQL for these cases.  Since the collection table is linked to many
other tables and the access table is for all types of objects, there
is difficulty in defining it to avoid 'cycles' and possible infinite
loops in generating the SQL.  But the specific-query is available for
cases like these.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@5072 ea137bd6-aa86-47b6-b90d-b101d9f19a90