Extensions so that 'irodssetup' can optionally use system-installed Postgresql and ODBC (i.e. installed via a package manager).
For ODBC, 'irodssetup' will notice that a key system ODBC file exists
(/usr/lib/odbc/psqlodbca.so) and if it does, ask whether the user wants
to use this (recommending so).

For Postgres, one enters the system location, for example:
  /usr/lib/postgresql/8.4

On Ubuntu (10.04.4 LTS), to install Postgres via apt-get, you would run:
  sudo apt-get install postgresql

For a system-installed Postgres, one needs to create the account to
use using the 'postgres' user.  For example, to create user
'schroeder':
  sudo -u postgres psql postgres
  create role schroeder with CREATEDB LOGIN;
  \password schroeder
  [enter password]
  \q

To install ODBC via apt-get:
  sudo apt-get install odbc-postgresql
  sudo apt-get install unixodbc-dev

The odbc-postgresql includes the /usr/lib/odbc/psqlodbca.so which is
the driver (which goes in the .odbc.ini file) and I believe
unixodbc-dev is needed for the include files for building irods.

It is also possible to use the system-installed ODBC and still
download and use various versions of Postgres.

In testing, I noticed that 'devtest' was somewhat slower when using
the system-installed postgres, which is probably because it's an older
version. On Ubuntu 10.04.4 LTS apt-get installs postgres 8.4 where we
download and use postgres 9.1.3.


git-svn-id: svn://irodssvn.ucsd.edu/trunk@5085 ea137bd6-aa86-47b6-b90d-b101d9f19a90