Fixes for Solaris. With these, it builds again and most of 'devtest' succeeds, including all of the ICAT tests.  See testiCommands.pl below for a description of the remaining problems.
config/config.mk.in:
For Solaris, use 'which gtar' to find the correct tar (GNU tar) to
use, as 'tar' will usually be the Solaris version which is not
compatible for 'ibun'.

server/icat/src/icatHighLevelRoutines.c:
Fixes for Solaris (C compiler).

server/icat/src/icatLowLevelOdbc.c:
Fixes for Solaris (C compiler).

lib/core/src/msParam.c:
Fixes for Solaris.  There is no strtof so use (float)strtod.

server/icat/src/convertSql.pl:
On Solaris, cpp may not be available so use 'gcc -E' instead to to
preprocess icatSysTables.sql.pp into icatSysTables.sql and
icatCoreTables.sql.pp into icatCoreTables.sql.

clients/icommands/test/testiCommands.pl:
Changed to use 'gtar' instead of 'tar' which fixes a few errors.

With that, four of the testiCommands.pl tests still fail:
95 - diff -r /var/tmp/testx /var/tmp/testx1/icmdtestx  ====> error code = 0
99 - diff -r /var/tmp/testx /var/tmp/icmdtestgz/icmdtestx  ====> error code = 0
104 - diff -r /var/tmp/testx /var/tmp/icmdtestbz2/icmdtestx  ====> error code = 0
145 - diff -r  /var/tmp/testx /var/tmp/testt  ====> error code = 0

The first one is just due to 'diff' differences; on solaris 'diff'
prints 'Common subdirectories' and lists the names so the check in
runCmd notices output and flags it as an error.  runCmd can check for
listed lines but not, as far as I can tell, with pattern matching.  To
get around this, you can't just pipe it thru 'grep -v Common' because
then the exit code is non-0 (from grep, not diff, I believe).  So a
work around would be to do diff|grep and then cat the file to check if
non-empty, altho then the listed command is 'cat'.

Even with the above change, the '99' diff fails, but this one
seems to be an actual problem with 'ibun'.

I'm not sure if 104 and 145 are just 'diff' problems or more
significant but thought I'd commit this work to this point.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@5100 ea137bd6-aa86-47b6-b90d-b101d9f19a90