More updates for Solaris for the testiCommands.pl script.  With this 'devtest' now succeeds on Solaris too, at least on merrimac.
Added an option in runCmd to ignore the return-code but still check
the stdout and added code to use this for SunOS and pipe the results
thru grep -v to remove the 'Common ...' lines.  This will give a
non-zero return code, but if there is text in the output it should
still notice the problem.

In the commit earlier today, I had indicated that the '99' test was
still getting an error, but with this change it seems to be OK.  As
far as I can tell the check in runCmd is working right.  I'm not sure
what was going wrong in the previous tests, but it seems OK now.


git-svn-id: svn://irodssvn.ucsd.edu/trunk@5101 ea137bd6-aa86-47b6-b90d-b101d9f19a90