Changes to improve performance for ichmod in instances where large numbers of access control items (ACLs) are being used (users/groups), for example at the iPlant Collaborative.
In chlModAccessControl and _modInheritance (called by it), changed the
SQL to no longer use a substr operation as that causes the indexes to
be inapplicable.  Instead, the functions create a path string, with
escaped (\) SQL special characters (_ and %) ending with /% that is
used via a 'like' clause (to find all sub-collections).

The psql 'explain' description indicates that this should perform
significantly better.  We have not yet been able to test it at scale
but it should perform at least as well, and appears to function
correctly.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@5164 ea137bd6-aa86-47b6-b90d-b101d9f19a90