Implementation of exclude patterns for ireg -C
There is a new option to ireg, '--exclude-from', which takes as an argument a file of pathname pattern matches evaluated by fnmatch(3). During a recursive registration, any pathname that matches one of the patterns will not be registered.

The --exclude-from file must be readable on the data server where the files will be registered, and the file format is one pattern per line. Lines starting with '#' are treated as comments.

Signed-off-by: Chris Smith <chris@distributedbio.com>

git-svn-id: svn://irodssvn.ucsd.edu/trunk@5228 ea137bd6-aa86-47b6-b90d-b101d9f19a90