Additions so that admins can insert multiple host names for certain resources (WOS).  New iadmin sub-options to 'modresc': host-add and host-rm to add or remove hosts from the list.
This is in support development being done at RENCI for DDN WOS
resources which have multiple network addresses.  The Agent will need
to parse these and choose an appropriate one to connect to (one that
is near).

ilsresc and 'iadmin lr' needed no changes to display what is now
optionally a command-separated list.

Tests were added to icatTest.pl to exercise the new SQL so that
'devtest' can still succeed.

Added host names are checked to see if they are valid DNS entries (as
is done for the 'iadmin mkresc' host namew).  If it is not valid, a
warning is returned but it is still used.

The column used is currently 250 characters.  For Postgres, the SQL
to increase this to 2700 is:
alter table R_RESC_MAIN alter column resc_net type varchar(2700);

See the iadmin help text for more.


git-svn-id: svn://irodssvn.ucsd.edu/trunk@5265 ea137bd6-aa86-47b6-b90d-b101d9f19a90