In general-query, handle the case where the client is requesting both uppercase (case insensitive) and numeric compares.
Avoid invalid SQL where the code was doing both upper() and and a
cast.  The cast is done on a per-column basis (users enter 'n<', 'n='
or 'n>') and the uppercase is for all 'where' columns (except now for
these).

Note that the user must use the 'n<' etc forms when using uppercase to
get this.  For example:

iquest uppercase "select DATA_NAME WHERE DATA_NAME like 'FX' and DATA_SIZE n< '1000'"

Now works correctly.

But:

iquest uppercase "select DATA_NAME WHERE DATA_NAME like 'FX' and DATA_SIZE < '1000'"

still gives a CAT_SQL_ERR.

This is because the code can't tell that the '<' in this case is
intended as a numeric compare.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@5278 ea137bd6-aa86-47b6-b90d-b101d9f19a90