Fix an apparent bug in msiChkHostAccessControl; set the error status so the Agent will exit if the check fails (user/host not configured).
This fix seems to be necessary to actually reject connections from
user-host combinations that are configured out in the
server/config/HostAccessControl file (see acChkHostAccessControl in
core.re).  Without it, a reject message is logged but the connection
is allowed.  Jean-Yves concurs that this may be an error, altho we
believe this had been working at some time.  I noticed this in testing
this morning related to a question on irods-chat.  



git-svn-id: svn://irodssvn.ucsd.edu/trunk@5291 ea137bd6-aa86-47b6-b90d-b101d9f19a90