In the i-command source files, change all the sprintfs to the snprintf form (with length limits) to avoid possible stack corruptions.
I don't think any of these were actual problems but the snprintf form
is safer.  There was a case in icd (fixed Monday and probably present
for years) in which a buffer could be written beyond the end causing a
stack corruption.  For this i-command code (particularly non-library
functions) this isn't too serious as it will just cause a client abort
which makes the problem apparent.  With snprint, the string will be
truncated which will cause problems later (but shouldn't happen
anyway); but snprintf is stll better form.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@5327 ea137bd6-aa86-47b6-b90d-b101d9f19a90