Implement a set of PAM auth extensions/configuration options as requested by the Swedish National Supercomputer Centre.
There is now a --ttl option on iinit for PAM mode to set the time to
live of the pam-derived irods password (within limits set by the
admin).

There's a config option, PAM_AUTH_NO_EXTEND, to disallow extensions to
the time that an existing pam-derived password is valid.  Without this
set they can still be extended but only before they expire
(previously, they were not usable but could be extended).

Pam-derived passwords are removed when expired (for all users) when
one is added.  This is especially useful when the PAM_AUTH_NO_EXTEND
is enabled as each user can create multiple passwords.

Multiple SQL changes were made in the icat code to handle these
changes.  This got fairly involved but I think it's better than
restructuring the ICAT password table at this point.

There is a change in a protocol message to include the TTL, but it is
only for iinit in PAM mode.

This was done in collaboration with Andreas Lindqvist of the Swedish
National Supercomputer Centre.

See notes and comments in the updated files for more information.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@5343 ea137bd6-aa86-47b6-b90d-b101d9f19a90