Improve the handling of an ICAT AVU metadata concurrency case.
When setting user-defined meta-data (AVUs), if two Agents try to
modify the the same metadata (which exists with different values in
R_META_MAIN), one of the chlSetAVUMetadata fails.  These changes allow
the agent to retry a second time.

Since this is in the case of "delete then add", the
chlDeleteAVUMetadata is called with noCommit=1, but the delete fails
(since the AVU was changed by the other agent), so in this case
(noCommit set) it should not do the rollback.  And then the caller,
chlSetAVUMetadata, can try a second time to modify the AVU.

Thanks to Thomas Ledoux of the National Library of France for these
improvements.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@5350 ea137bd6-aa86-47b6-b90d-b101d9f19a90