Changes so iRODS can be built with HPSS again (post 3.0).
It looks like one cannot link HPSS client libraries using g++, at
least on some platforms.  Since we changed to building with g++ in
iRODS 3.0, this may have been a problem since then.  We experimented
with various ways to try to link (using -fPIC and -shared) but that
created incompatibilities with other libraries.  I believe that an
HPSS bug report been submitted so it may be fixed in a future HPSS
version.

To work around this, we've modified platform.mk.template (platform.mk)
so that if HPSS is selected, gcc will be used instead of g++ (for both
compiling and linking).

The change in config.mk.in (config.mk) is needed when building on
64-bit hosts.  I'm not sure why there was a line there to unset
ADDR_64BITS= as that variable is set (or not) elsewhere (near the top
in config.mk).  But it was a problem for us.

The mod in hpssFileDriver.h is enabled with g++ so it is not currently
active but will be needed if/when g++ can be used.  The need for this
was discovered in some of our testing so I'm putting it in now.

The mod in hpssFileDriver.c is also needed for g++, but not with gcc,
but is still a useful change.

This is for HPSS 7.3.x.  For HPSS 7.4 it is believed that the iRODS
driver will need to change to using the HPSS PIO API as the current
hpss_writelist API is deprecated.

This work was done in collaboration with Jeffrey Cox at Indiana
University.


git-svn-id: svn://irodssvn.ucsd.edu/trunk@5397 ea137bd6-aa86-47b6-b90d-b101d9f19a90