The following changes allow iRODS to be built with Boost enabled again on Linux systems using the system-installed Boost instead of the version included in the iRODS release/repository.  Next steps will be to remove the boost library source code and support from the repository.
Previously, in collaboration with RENCI, supoort for Boost was added
in iRODS 3.0, primarily to make the Windows version immediately
available.  For Linux/Unix, Boost was a build-time option but has not
been working recently (3.2 and probably 3.1).  The DICE continuous
test system (tinderbox) builds with Boost enabled daily, but the test
was inadvertently not failing when the (re) make failed (this was
fixed a few weeks ago).  Windows support also became a lower priority
effort (altho DICE and RENCI do still support it, and this commit is
in support of that effort).

Additional changes made in the iRODS code post-3.0 make the current
iRODS code incompatible with the version of Boost in the iRODS
repository.  But boost can now be easily installed as a system library
using standard package managers and the version available with recent
Linux releases is sufficient.  This is being tested under Ubuntu
12.04.2 which includes Boost 1.46.1.



git-svn-id: svn://irodssvn.ucsd.edu/trunk@5417 ea137bd6-aa86-47b6-b90d-b101d9f19a90